親子関係の追加

+++ Project/Game/Editer/BossAttackEditer.cpp
@@ -289,7 +289,7 @@ void BossAttackEditer::SelectAttack() {
 	
 	if (ImGui::BeginCombo(".json##", &selectFileName_[0], ImGuiComboFlags_HeightLargest)) {
 		for (int i = 0; i < attackFileNames_.size(); i++) {
-			const bool isSelected = (selectIndex_ == i);
+			const bool isSelected = ((int)selectIndex_ == i);
 			if (ImGui::Selectable(attackFileNames_[selectIndex_].c_str(), isSelected)) {
 				selectIndex_ = i;
 			}
